split up config_metplus into 3 files so the utils are easier to navigate
