+++ metplus/util/__init__.py
@@ -4,8 +4,10 @@ from .string_manip import *
 from .system_util import *
 from .time_util import *
 from .string_template_substitution import *
-from .doc_util import *
+from .config_util import *
 from .config_metplus import *
+from .config_validate import *
+from .doc_util import *
 from .run_util import *
 from .met_config import *
 from .time_looping import *
